pyproject.toml
[tool.poetry]
name = "cloak"
version = "0.1.0"
description = ""
authors = ["vcokltfre <vcokltfre@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
ormar = "^0.12.1"
asyncpg = "^0.27.0"
alembic = "^1.11.0"
psycopg2-binary = "^2.9.6"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
taskipy = "^1.10.4"

[tool.taskipy.tasks]
format = "black . && isort --profile black ."

db-migrate = "alembic revision --autogenerate -m"
db-upgrade = "alembic upgrade head"
start = "task db-upgrade && uvicorn src.api.__main__:app --reload --host 0.0.0.0 --port 8080"

dev-db-migrate = "python3 scripts/env.py alembic revision --autogenerate -m"
dev-db-upgrade = "python3 scripts/env.py alembic upgrade head"
dev-start = "task dev-db-upgrade && uvicorn src.api.__main__:app --reload --host 127.0.0.1 --port 8080"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
